+++ test/bfp_tests/src/bfp/complex/test_bfp_complex_energy.c
@@ -45,7 +45,7 @@ TEST(bfp_complex_energy, bfp_complex_s16_energy)
     for(int r = 0; r < REPS; r++){
         setExtraInfo_RS(r, seed);
 
-        bfp_complex_s16_init(&B, B_real, B_imag, 
+        bfp_complex_s16_init(&B, B_real, B_imag,
             pseudo_rand_int(&seed, -100, 100),
             pseudo_rand_uint(&seed, 0, MAX_LEN+1), 0);
 
@@ -53,7 +53,7 @@ TEST(bfp_complex_energy, bfp_complex_s16_energy)
 
         double expected_f = 0.0;
 
-        for(int i = 0; i < B.length; i++){
+        for(unsigned int i = 0; i < B.length; i++){
             B.real[i] = pseudo_rand_int16(&seed) >> B.hr;
             B.imag[i] = pseudo_rand_int16(&seed) >> B.hr;
 
@@ -70,7 +70,7 @@ TEST(bfp_complex_energy, bfp_complex_s16_energy)
 
         TEST_ASSERT_EQUAL(0, conv_error);
 
-        unsigned threshold = ceil(log2(B.length));
+        unsigned threshold = (unsigned) ceil(log2(B.length));
 
         TEST_ASSERT_EQUAL_INT64_MESSAGE(expected, result.mant, "");
     }
@@ -99,7 +99,7 @@ TEST(bfp_complex_energy, bfp_complex_s32_energy)
 
         double expected_f = 0.0;
 
-        for(int i = 0; i < B.length; i++){
+        for(unsigned int i = 0; i < B.length; i++){
             B.data[i].re = pseudo_rand_int32(&seed) >> B.hr;
             B.data[i].im = pseudo_rand_int32(&seed) >> B.hr;
 
@@ -109,7 +109,7 @@ TEST(bfp_complex_energy, bfp_complex_s32_energy)
 
         // if(r == 11){
         //   printf("B.exp = %d\n", B.exp);
-        //   for(int k = 0; k < B.length; k++){
+        //   for(unsigned int k = 0; k < B.length; k++){
         //     printf("(%ld, %ld), ", B.data[k].re, B.data[k].im);
         //   }
         //   printf("\n\n");
@@ -139,7 +139,7 @@ TEST(bfp_complex_energy, bfp_complex_s32_energy)
 
         // printf("(%d, %u, %ld), ", B.length, B.hr, (int32_t)(expected - result.mant));
 
-        // If left-shifting by at least 15 bits, the result should be exact (because loss only 
+        // If left-shifting by at least 15 bits, the result should be exact (because loss only
         // happens when b_shr is positive, and on the implicit 30-bit right-shift, and if b_shr
         // is <= -15 the 30-bit right-shift can only shift out zeros).
         if( b_shr <= -15 )
